fileOpen.c -> file.c, file.h
